requirements.txt
Flask==3.0.3
PyYAML==6.0.2
python-dotenv==1.0.0
supabase==2.3.4
